cobra: vm: Get CPU usage in percentage
In order to do anything with CPU times, and in order to
take CPU into consideration, we arbitrarily have converted
the CPU time into average CPU usage in percentage to
perform operations.

This is a dumb hack to get it to work; the data is not always
precise - we should snapshot the data for this to work better.
It uses the delay specified in vm.ini to capture data - it
is always good to specify the delay there to 1 - which is 1
second.

Now that we have no use for system_time, and user_time, remove
the unused variables.

The formula for CPU usage was referred from
https://github.com/virt-manager/virt-manager/blob/bb1afaba29019605a240a57d6b3ca8eb36341d9b/virtManager/lib/statsmanager.py.

Sincere thank you to RedHat for virt-manager.

TODO: Use snapshots, and use proper classes to get the work done.
Signed-off-by: Gagan Malvi <malvigagan@gmail.com>